[WIP] Implementation of n-d TicTacToe and minimax
Includes an interface for zero sum games.
Minimax implementation that can use any deterministic sequential
zero sum game

TicTacToe with an arbitrarially sized game board and arbitrary lenght
win condition is implemented.

Maven build system used for this project because I expect it to grow
to a sufficiant size that a real build system is required

Heuristic evaluation function is not finished.  Minimax move stack needs
to be implemeted